Rename elevatorarm folder/package to wrist
