+++ src/main/java/frc/team4909/robot/subsystems/wrist/commands/ElevatorArmOperatorControl.java
@@ -1,4 +1,4 @@
-package frc.team4909.robot.subsystems.elevatorarm.commands;
+package frc.team4909.robot.subsystems.wrist.commands;
 
 import frc.team4909.robot.Robot;
 import frc.team4909.robot.RobotConstants;
