+++ src/main/java/frc/team4909/robot/subsystems/wrist/commands/SetWristAngle.java
@@ -1,4 +1,4 @@
-package frc.team4909.robot.subsystems.elevatorarm.commands;
+package frc.team4909.robot.subsystems.wrist.commands;
 
 import edu.wpi.first.wpilibj.command.InstantCommand;
 import frc.team4909.robot.Robot;
